requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-jazzmin==3.0.1
django-modeltranslation==0.19.19
django-phonenumber-field==8.4.0
dotenv==0.9.9
idna==3.11
phonenumbers==9.0.22
pillow==12.1.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
